added ntd to crosswalk stuff, rm all fillna for dir id and reran
