+++ gtfs_funnel/schedule_stats_by_route_direction.py
@@ -21,18 +21,18 @@ def cardinal_direction_for_route_direction(analysis_date:str, dict_inputs:dict):
     stop_times_gdf = pd.read_parquet(
         f"{RT_SCHED_GCS}{STOP_TIMES_FILE}_{analysis_date}.parquet",
         filters=[[("stop_primary_direction", "!=", "Unknown")]
-        ])
+    ])
     
     trip_scheduled_col = [
-    "route_id",
-    "trip_instance_key",
-    "gtfs_dataset_key",
-    "shape_array_key",
-    "direction_id",
-    "route_long_name",
-    "route_short_name",
-    "route_desc",
-    "name"
+        "route_id",
+        "trip_instance_key",
+        "gtfs_dataset_key",
+        "shape_array_key",
+        "direction_id",
+        "route_long_name",
+        "route_short_name",
+        "route_desc",
+        "name"
     ]
         
     trips_df = helpers.import_scheduled_trips(analysis_date, 
@@ -46,9 +46,6 @@ def cardinal_direction_for_route_direction(analysis_date:str, dict_inputs:dict):
                   "shape_array_key"]
     
     stop_times_with_trip = pd.merge(stop_times_gdf, trips_df, on = merge_cols)
-    
-    # Fill in missing direction id with 0, per our usual practice.
-    # stop_times_with_trip.direction_id = stop_times_with_trip.direction_id.fillna(0)
     
     main_cols = [
         "route_id",
@@ -127,9 +124,7 @@ def assemble_scheduled_trip_metrics(
         on = "trip_instance_key",
         how = "inner"
     )
-    
-    #median_stop_meters_df.direction_id = median_stop_meters_df.direction_id.fillna(0)
-   
+
     return median_stop_meters_df
